backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
redis
python-dotenv
openai
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
transformers
scikit-learn
numpy
asyncio-mqtt
aiofiles
pytest
httpx
PyPDF2
python-docx
python-multipart
asyncpg
pydantic-settings
boto3
tiktoken
